+++ CHANGELOG.md
@@ -157,3 +157,4 @@
 - [2026-04-06 12:21] chore(core): update dependencies and cleanup unused imports in core
 - [2026-04-06 12:23] feat(ui): implement responsive ui view component
 - [2026-04-06 12:32] feat(storage): add bi-directional event bus handler for storage
+- [2026-04-06 13:17] feat(i18n): implement responsive i18n view component